Added histogram function and empty file test

+++ opc.py
@@ -6,6 +6,7 @@ import time
 import struct
 import datetime
 import sys
+import os.path
 
 # Init
 def initOPC(ser):
@@ -46,6 +47,9 @@ def fanOn(ser):
 	print(nl)
         time.sleep(.1)
 
+def combine_bytes(LSB, MSB):
+	return (MSB << 8) | LSB
+
 def getHist(ser):
 	ser.write(bytearray([0x61,0x30]))
 	nl=ser.read(2)
@@ -54,11 +58,31 @@ def getHist(ser):
 	br = bytearray([0x61])
 	for i in range(0,62):
 		br.append(0x30)
-	print(len(br))
 	ser.write(br)
-	ans=bytearray(ser.read(63))
-	print(len(ans))
-	print(ans[1:63])
+	ans=bytearray(ser.read(1))
+	ans=bytearray(ser.read(62))
+	data={}
+	data['Bin 0'] = combine_bytes(ans[0],ans[1])
+	data['Bin 1'] = combine_bytes(ans[2],ans[3])
+	data['Bin 2'] = combine_bytes(ans[4],ans[5])
+	data['Bin 3'] = combine_bytes(ans[6],ans[7])
+	data['Bin 4'] = combine_bytes(ans[8],ans[9])
+	data['Bin 5'] = combine_bytes(ans[10],ans[11])
+	data['Bin 6'] = combine_bytes(ans[12],ans[13])
+	data['Bin 7'] = combine_bytes(ans[14],ans[15])
+	data['Bin 8'] = combine_bytes(ans[16],ans[17])
+	data['Bin 9'] = combine_bytes(ans[18],ans[19])
+	data['Bin 10'] = combine_bytes(ans[20],ans[21])
+	data['Bin 11'] = combine_bytes(ans[22],ans[23])
+	data['Bin 12'] = combine_bytes(ans[24],ans[25])
+	data['Bin 13'] = combine_bytes(ans[26],ans[27])
+	data['Bin 14'] = combine_bytes(ans[28],ans[29])
+	data['Bin 15'] = combine_bytes(ans[30],ans[30])
+	data['period'] = struct.unpack('f',bytes(ans[44:48]))[0]
+	data['pm1'] = struct.unpack('f',bytes(ans[50:54]))[0]
+	data['pm2'] = struct.unpack('f',bytes(ans[54:58]))[0]
+	data['pm10'] = struct.unpack('f',bytes(ans[58:]))[0]
+	return(data)
 	
 # Retrieve data
 def getData(ser):
@@ -106,15 +130,19 @@ if __name__ == "__main__":
 	time.sleep(5)	
 
 	print("Opening Output File:")
-	f=open(ofile,'w+')
-	print("time,pm1,pm2,pm10",file=f)
+	if(not os.path.isfile(ofile)):
+		f=open(ofile,'w+')
+		print("time,b0,b1,b2,b3,b4,b5,b6,b7,b8,b9,b10,b11,b12,b13,b14,b15,period,pm1,pm2,pm10",file=f)
+	else:
+		f=open(ofile,'a')
 	print("Looping:")
 	for i in range(0,4320):
-		t=getData(ser)
+		t=getHist(ser)
 		ts = time.time()
 		tnow = datetime.datetime.fromtimestamp(ts).strftime('%Y-%m-%d %H:%M:%S')
-		print(tnow + "," + str(t[0]) + "," + str(t[1]) + "," + str(t[2]), file=f)
-		print(tnow + "," + str(t[0]) + "," + str(t[1]) + "," + str(t[2]))
+		data=t
+		print(tnow + "," + str(data['Bin 0']) + ","  + str(data['Bin 1']) + ","  + str(data['Bin 2']) + ","  + str(data['Bin 3']) + ","  + str(data['Bin 4']) + ","  + str(data['Bin 5']) + ","  + str(data['Bin 6']) + ","  + str(data['Bin 7']) + ","  + str(data['Bin 8']) + ","  + str(data['Bin 9']) + ","  + str(data['Bin 10']) + ","  + str(data['Bin 11']) + ","  + str(data['Bin 12']) + ","  + str(data['Bin 13']) + ","  + str(data['Bin 14']) + ","  + str(data['Bin 15']) + ","  + str(data['period']) + ","  + str(data['pm1']) + ","  + str(data['pm2']) + ","  + str(data['pm10']) , file=f)
+		print(tnow + "," + str(data['pm1']) + "," + str(data['period']) + "," + str(data['Bin 15']))
 		f.flush()
 		time.sleep(59)
